security check for "annulation"

+++ app/Http/Controllers/AnnulationController.php
@@ -15,13 +15,27 @@ class AnnulationController extends Controller
     //
     public function show($token){
         $title = Config::get('configuration.title');
-        return view('annulation', compact('token', 'title'));   
+        $reservationAnnulation = Reservation::where('uniqueId', $token)->get();
+
+        if(count($reservationAnnulation)>0){
+            return view('annulation', compact('token', 'title', 'reservationAnnulation')); 
+        }
+        else{
+            return redirect('/')->with('error','Aucune réservation trouvée.');
+        } 
     }
 
     public function destroy($token)
     {
-        Reservation::where('uniqueId', $token )->delete();
+        $reservationAnnulation = Reservation::where('uniqueId', $token)->get();
+        if(count($reservationAnnulation)>0){
+            Reservation::where('uniqueId', $token )->delete();
 
-        return redirect('/')->with('success','Votre réservation a bien été annulée');
+            return redirect('/')->with('success','Votre réservation a bien été annulée');
+        }
+        else{
+            return redirect('/')->with('error','Aucune réservation trouvée.');
+        } 
+        
     }
 }
